[HW][Comb][ExportVerilog] Fix HWIntegerType's C++ type to allow hw::IntType.
Previously, the dialect type HWIntegerType had a declared C++ type of
mlir::IntegerType. This was fine in the past, when HWIntegerType
was just a lightweight wrapper around the MLIR type. However, this
gradually became incorrect when we created !hw.int for holding
hardware integers with parametric width and as we started using !hw.int
in existing ops that were declared to have HWIntegerType operands and
results.

Although it was technically incorrect, this didn't cause issues because:

1. Parametric !hw.int isn't really fully supported, and most existing
   passes assume we have statically-known widths.
2. Because the C++ type was incorrect, most of the code implicictly
   converted the operand/result types to mlir::IntegerType, which would
   work fine as long as the runtime type was indeed mlir::IntegerType.

2) is particularly pernicious, since it masked places where we should
have been explicitly casting values to mlir::IntegerType, since the
compiler would assume that it should implicitly convert to
mlir::IntegerType without considering the possibility that we actually
have an hw::IntType.

We need to actually fix this now that upstream MLIR introduced a
TypedValue type, since TypedValue will perform a runtime assertion to
check that the wrapped value is indeed of the correct type.
Specifically, ops that have a dialect type of HWIntegerType would
declare a C++ type of mlir::IntegerType, and if the op _actually_ had a
type of hw::Int, the new TypedValue would fail the runtime type check.

This commit changes HWIntegerType's C++ type to be
TypeOr<mlir::IntegerType, hw::IntType>, which correct accounts for both
possible cases, and it allows for the TypedValue runtime type checks to
succeed.

This change has the side effect of forcing us to explicitly cast most
ops' results to mlir::IntegerType in places where we currently assume
that the type is a non-parametric hardware integer. Ideally, we should
be checking whether an op result is parametric before we run folds that
rely on having a statically-known width, but putting in the explicit
casts preserves the existing behavior while at least making it more
explicit that we need an integer of statically-known width.