[FIRRTL] Introduce const types
Const types are versions of FIRRTL hardware types that can only be assigned
compile-time constant values or values of other const types. Currently only
integer and aggregates of integer types can be const.

This commit introduces const support in the FIRRTL type system but does not
yet support using const types in operations except for connects. Parsing and
emitting const in fir files will be supported in a future pr.