[HWMemSimImpl] Add Memory File Initialization
Add support to HWMemSimImpl for doing file-based memory initialization.
This comes in two flavors:

  1. Inline initialization
  2. Bind-based initialization

In (1), a ReadMemOp is used to initialize the memory from a file.
In (2), a separate module is created that initializes the memory via
ReadMemOp using an XMRRefOp.  This new module is then instantiated in
the memory module via a bind.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>